move test to recursion compiler circuit
